[Kinetics] Move PlogRate to its own header file
